requirements.txt
streamlit
pytesseract
PyMuPDF
python-dotenv
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
imagehash
langchain-google-genai
langchain-core
Pillow
torch
langchain
google-cloud-storage
google-auth
datasets
transformers
safetensors